inst/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
fastapi==0.104.1
Flask==3.1.2
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.2
jiter==0.9.0
MarkupSafe==3.0.2
multidict==6.4.4
openai==1.37.0
propcache==0.3.1
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
PyJWT==2.10.1
pyTelegramBotAPI==4.27.0
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
rsa==4.9.1
sniffio==1.3.1
starlette==0.27.0
telebot==0.0.5
Telethon==1.40.0
tqdm==4.67.1
twilio==8.10.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.23.2
websockets==12.0
Werkzeug==3.1.4
yarl==1.20.0
